pyproject.toml
[project]
name = "local-police-events"
version = "0.1.0"
description = "Calendar feed service for UK police neighbourhood events"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "duckdb>=1.1.0",
    "icalendar>=6.0.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "pyproj>=3.7.0",
    "cachetools>=5.3.0",
]
